refactor(anni): further abstract abstract annotation
